+++ packages/expo-notifications/src/Notifications.types.ts
@@ -526,6 +526,10 @@ export type NotificationContentAndroid = {
    * The pattern with which to vibrate.
    */
   vibrationPattern?: number[];
+  /**
+   * An identifier used to group related notifications together.
+   */
+  threadIdentifier?: string | null;
 };
 
 /**
@@ -611,8 +615,11 @@ export type NotificationContentInput = {
    */
   categoryIdentifier?: string;
   /**
-   * An identifier used to group related notifications together in the notification center.
-   * @platform ios
+   * An identifier used to group related notifications together.
+   *
+   * On Android: the notification's group key, which enables notification stacking.
+   *
+   * On iOS: `threadIdentifier`, which groups notifications in the notification center.
    */
   threadIdentifier?: string;
   /**